Initial development on a generic udp dns scanning module
This module is not finished. Current status is it correctly
generates, validates, and processes packets. It does not
correctly parse response RR's yet (compression ugh).

This is a rough skeleton and first pass at a complete rebuild of
the deprecated module_udp_dns module. This new module supports
10 question types (e.g. A, AAAA, ANY, MX, RRSIG...) and will return
completely parsed DNS responses. Since DNS responses are not flat
the fieldsets returned are nested.